UserInput -> GameInput : add features for player screen
